% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_freq_table.R
\name{compute_freq_table}
\alias{compute_freq_table}
\title{Title}
\usage{
compute_freq_table(data, nlev, j, parentnodes, lookup = NULL)
}
\arguments{
\item{data}{}

\item{nlev}{}

\item{j}{}

\item{parentnodes}{}

\item{lookup}{}
}
\description{
Title
}
