% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_local_bdeu_score.R
\name{compute_local_bdeu_score}
\alias{compute_local_bdeu_score}
\title{Title}
\usage{
compute_local_bdeu_score(
  counts,
  ess,
  r = ncol(counts),
  q = nrow(counts),
  s = 1
)
}
\arguments{
\item{counts}{}

\item{ess}{}

\item{r}{}

\item{q}{}

\item{s}{}
}
\description{
Title
}
