% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prec_recall.R
\name{compute_prec_recall}
\alias{compute_prec_recall}
\title{Title}
\usage{
compute_prec_recall(x, y, method)
}
\arguments{
\item{x}{}

\item{y}{}

\item{method}{}
}
\description{
Title
}
\examples{

x <- c(0, 1, 1)
y <- c(0, 0, 1)
compare_prec_recall(x, y, niter = 20)

x <- c(1, 1, 3, 3, 3)
y <- c(0, 1, 1, 1, 0)
compare_prec_recall(x, y, niter = 20)

x <- c(rep(0, 20), 1:80)
y <- c(runif(20) < .2, runif(80) < c(1:80)/100)
compare_prec_recall(x, y, niter = 20)



}
\keyword{internal}
