% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_scorepar.R
\name{define_scorepar}
\alias{define_scorepar}
\title{Title}
\usage{
define_scorepar(
  data,
  nlev,
  ess = 1,
  edgepf = 2,
  local_struct = NULL,
  regular = FALSE,
  lookup = NULL
)
}
\arguments{
\item{data}{(matrix)
each column is assumed to be outcomes of a categorical variable, with
possible outcomes \verb{0, ..., K-1.}}

\item{nlev}{(integer vector)
cardinality of each variable.}

\item{ess}{(numeric constant)
equivalent sample size}

\item{edgepf}{(numeric constant)
a factor \code{-log(edgepf)*length(parentnodes)} is added to the marginal likelihood score.}

\item{local_struct, }{regular
parameters controlling the local-structure-optimization routine. See \code{\link[=optimize_partition]{optimize_partition()}}.}

\item{lookup}{(environment)
an environment in which the optimized structures are saved, initiated by rlang:::new_environment().
If \code{NULL} the structures are not  saved.}
}
\value{
an object of class \link[BiDAG:scoreparameters]{BiDAG::scoreparameters}
}
\description{
Title
}
