% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famscore_bdeu_byrow.R
\name{famscore_bdeu_1row}
\alias{famscore_bdeu_1row}
\title{Compute BDEu-score for each parent config}
\usage{
famscore_bdeu_1row(x, ess, r = length(x), q = 1, s = 1)
}
\arguments{
\item{x}{a matrix with observed counts}

\item{s}{scaling factor for the hyperparam \code{alpha}}
}
\value{
an array with score for each parent config
}
\description{
Compute BDEu-score for each parent config
}
