% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_search_space.R
\name{init_search_space}
\alias{init_search_space}
\title{Initialize search space}
\usage{
init_search_space(
  scorepar,
  algo_init,
  hardlimit,
  maxp = hardlimit,
  alpha = 0.05,
  verbose
)
}
\arguments{
\item{scorepar}{(object of class \link[BiDAG:scoreparameters]{BiDAG::scoreparameters})
See \link{define_scorepar}.}

\item{algo_init}{(character)
Name of optimization routine for local structure.
See \link{optimize_partition}.}

\item{hardlimit}{(integer)
Maximal number of parents allowed. See \link[BiDAG:learnBN]{BiDAG::learnBN}.}

\item{maxp}{(integer)
Maximal number of parents in CPDAG learned by \link[bnlearn:hc]{bnlearn::hc}.
Ignored if not \code{algo_int=="hc"} or \code{algo_init == "hcskel"}.}

\item{alpha}{(numeric)
Significance level for the conditional independencies test in \link[pcalg:pc]{pcalg::pc}.
Ignored if not \code{algo_int=="pc"} or \code{algo_init == "pcskel"}.}

\item{verbose}{(logical)}
}
\value{
a \code{n-by-n} adjacency matrix
}
\description{
Wrapper around structural learning routines (\link[bnlearn:hc]{bnlearn::hc} and \link[pcalg:pc]{pcalg::pc}),
intended to learn a start-space for the \link[Bidag:iterativeMCMC]{Bidag::iterativeMCMC} procedure.
Forces the maximal number of parents (\code{hardlimit}) to be respected,
by re-running the structure learning procedure with stricter add-edge-policies
(\code{maxp} and \code{alpha}, respectively) until a structure that satifies the \code{hardlimit}
criteria is inferred.
}
\seealso{
\link{sample_dags}
}
