% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_to_partition.R
\name{labels_to_partition}
\alias{labels_to_partition}
\title{Title}
\usage{
labels_to_partition(
  labels,
  levels,
  nlev = lengths(levels),
  type = "outcome_vectors"
)
}
\arguments{
\item{levels}{(list)
list with levels of each parent}

\item{nlev}{(integer vector)}

\item{type}{(character)
how to read labels. Either \code{"outcome_vectors"} (default) or \code{"enumerated"}.}

\item{label}{(list)
list with labels on edge from each parent to their common child.
Labels specified as a matrix with each row the joint outcomes of co-parents,
or as a vector that enumerates these joint outcomes (the rowindex of the CPT
table starting from 0). See example.}
}
\description{
Title
}
\examples{
levels <- list(0:1, 0:2, 0:3)
label <- vector("list", length(nlev))
label[[1]] <- rbind(c(NA, 0), c(0, 1))
part1 <- labels_to_partition(label, levels, type = "outcome_vectors")

label[[1]] <- c(0, 2, 4, 6)
label[[2]] <- c(0)
part2 <- labels_to_partition(label, levels, type = "enumerated")

cbind(expand.grid(levels),  get_partitions(part1), get_partitions(part2))

}
