% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_partition_regular.R
\name{make_partition_regular}
\alias{make_partition_regular}
\title{Make a partition regular}
\usage{
make_partition_regular(P, nlev)
}
\arguments{
\item{P}{}

\item{nlev}{}
}
\description{
If the partition implies a conditional independence, return a finer partitioning
that does not. Specifically, if all outcomes of one or more variables are included
in all parts of the partition, divide each part by the outcomes of these variables.
}
\examples{
nlev <- c(2, 2)
P <- list(c(0, 2), c(1, 3))
make_regular_partition(P, nlev)
}
