% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regular.R
\name{make_regular}
\alias{make_regular}
\title{Make a partition regular}
\usage{
make_regular(P, nlev)
}
\arguments{
\item{P}{(list) a partition of the}

\item{nlev}{(integer vector) the cardinality of the variables}
}
\description{
If the partition implies any conditional independencies, return a finer partitioning
that does not, by splitting all parts by the outcomes of the relevant variable(s).
}
\examples{
nlev <- c(2, 2)

P <- list(0:3)
make_regular(P, nlev)

P <- list(c(0, 2), c(1, 3))
make_regular(P, nlev)

P <- list(0, 1:3)
make_regular(P, nlev)
}
\seealso{
\code{\link[=is_regular]{is_regular()}}
}
