% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partition.R
\name{partition}
\alias{partition}
\alias{new_partition}
\alias{get_parts}
\title{Class: partition}
\usage{
new_partition(x, nlev)

get_parts(P)
}
\arguments{
\item{x}{(list)
a list with}

\item{nlev}{integer vector:
the cardinality of each variable}

\item{P}{an object of class \code{partition}}
}
\value{
an object of class \code{partition}

\itemize{
\item \verb{get_parts:} a vector with that assigns each outcome to a part
}
}
\description{
An object of class \code{partition} is a list that defines the partition of the
outcome space of a set of categorical variables.
}
\examples{
# generate a new partition
nlev <- c(2, 2)
P <- new_partition(list(c(1, 3), 2, 0), nlev)
P
# list which part each outcome belongs to
get_parts(P)
}
