% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_bn.R
\name{rand_bn}
\alias{rand_bn}
\title{Title}
\usage{
rand_bn(dag, partitions = NULL, nlev = rep(2, ncol(dag)), alpha = 1)
}
\arguments{
\item{dag}{}

\item{nlev}{}

\item{alpha}{}

\item{labels}{}
}
\description{
Title
}
\examples{
dag <- rbind(L  = c(0, 1, 1, 0, 1),
             Z1 = c(0, 0, 0, 1, 0),
             Z2 = c(0, 0, 0, 1, 0),
             X  = c(0, 0, 0, 0, 1),
             Y  = rep(0, 5))
colnames(dag) <- rownames(dag)
nlev <- rep(2, ncol(dag))

# label structure
labels <- vector("list", 5)
labels[[4]] <- list(0, 0)
labels[[5]] <- list(0, NULL)

rand_bn(dag, nlev, alpha = 1, labels = labels)
}
