% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dags.R
\name{sample_dags}
\alias{sample_dags}
\title{Title}
\usage{
sample_dags(
  scorepar,
  algo_init = "pcskel",
  algo_sample = "order",
  hardlimit = 5,
  verbose = F
)
}
\arguments{
\item{scorepar}{(object of class \link[BiDAG:scoreparameters]{BiDAG::scoreparameters})
See \link{define_scorepar}.}

\item{algo_init}{(character)
Name of optimization routine for local structure.
See \link{optimize_partition}.}

\item{algo_sample}{(character)
Name of MCMC implementation. Either \code{partition} or \code{order}.
See \link[BiDAG:sampleBN]{BiDAG::sampleBN}.}

\item{hardlimit}{(integer)
Maximal number of parents allowed. See \link[BiDAG:learnBN]{BiDAG::learnBN}.}

\item{verbose}{(logical)}
}
\description{
Title
}
