% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_from_lookup.R
\name{score_from_lookup}
\alias{score_from_lookup}
\title{Compute local score in a labeled DAG from data}
\usage{
score_from_lookup(
  data,
  levels,
  nlev = lengths(levels),
  j,
  parentnodes,
  ess,
  method,
  regular,
  lookup
)
}
\arguments{
\item{data}{}

\item{nlev}{}

\item{j}{node to be scored}

\item{parentnodes}{}

\item{method}{}

\item{table}{}
}
\value{
score of node \code{j} given \code{parentnodes} and optimal CSI-structure
}
\description{
Compute sample counts over the relevant subset of variables and compute score.
If more than one parent return the score of the optimal local CSI-structure.
}
