% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_scorepar.R
\name{define_scorepar}
\alias{define_scorepar}
\title{Title}
\usage{
define_scorepar(
  data,
  nlev,
  ess = 1,
  edgepf = 2,
  local_struct = NULL,
  regular = FALSE,
  lookup = NULL
)
}
\arguments{
\item{data}{(matrix)
Each column is assumed to be outcomes of a categorical variable, with
possible outcomes \verb{0, ..., K-1.}}

\item{nlev}{(integer vector)
Cardinality of each variable.}

\item{ess}{}

\item{edgepf}{}

\item{local_struct}{}

\item{regular}{}

\item{lookup}{}
}
\description{
Title
}
