% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_CSI_consistent_partition.R
\name{is_CSI_consistent}
\alias{is_CSI_consistent}
\alias{is_CSI_consistent_part}
\title{Check if a partition is CSI consistent}
\usage{
is_CSI_consistent(
  P,
  levels,
  nlev = lengths(levels),
  stride = c(1, cumprod(nlev[-length(nlev)]))
)

is_CSI_consistent_part(p, levels, nlev, stride)
}
\arguments{
\item{P}{(list of integer vectors)
a partition of the outcome space defined by \code{levels}.
enumerating the joint outcomes from \code{0} to \code{prod(lengths(nlev))-1}.}

\item{levels}{(list of integer vectors)
a list with outcomes of each variable.}

\item{stride}{(integer vector)
the stride of each variable, corresponding to levels.}

\item{p}{(integer vector)
a part of \code{P}, i.e. a subset of joint outcomes.}
}
\value{
a logical constant
}
\description{
Check if a partition is CSI consistent
}
\examples{
levels <- list(0:1, 0:1)

# CSI-consistent
P <- list(c(0, 1, 3), 2)
is_CSI_consistent(P, levels)

# not CSI-consistent partition
P <- list(c(0, 3), 2, 1)
is_CSI_consistent(P, levels)

P <- list(0, 1:2, 3)
is_CSI_consistent(P, levels)


# CSI-consistent, but not regular
P <- list(c(0, 1), c(2, 3))
is_CSI_consistent(P, levels)
is_regular(P, lengths(levels))
}
