% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_partition.R,
%   R/optimize_partition_ldag.R, R/optimize_partition_tree.R
\name{optimize_partition}
\alias{optimize_partition}
\alias{optimize_partition_from_cpt}
\alias{optimize_partition_ldag}
\alias{optimize_partition_tree}
\title{Optimize partition over parent outcome space}
\usage{
optimize_partition(
  counts,
  levels,
  ess,
  method,
  make_regular = F,
  verbose = FALSE
)

optimize_partition_from_cpt(
  cpt,
  ess,
  method,
  make_regular = T,
  return_cpt = TRUE
)

optimize_partition_ldag(
  counts,
  levels,
  ess,
  P = as.list(1:nrow(counts) - 1),
  labels = rep(list(integer()), length(levels)),
  verbose = FALSE
)

optimize_partition_tree(counts, levels, ess, verbose = verbose)
}
\arguments{
\item{counts}{(integer matrix)
a frequency table.}

\item{levels}{(list of integer vectors)
levels of each conditioning variable in \code{counts},
such that \code{expand.grid(levels)} gives the joint configurations corresponding
to each row in \code{counts}.}

\item{ess}{(numeric)
imaginary sample size for the bdeu-prior.}

\item{method}{(character)}

\item{make_regular}{(logical)
if \code{FALSE} (default) the optimized partition is returned, also if it implies
conditional independencies (is not "regular"). If \code{TRUE} each part of the
optimized partition is divided by the outcomes of the relevant variables.}

\item{verbose}{(logical)}
}
\value{
a list with named elements:
\itemize{
\item \code{partition}: the partition implied by the tree
\item \code{scores}: each part's contribution to the local score
\item additional output from the different optimization procedures.n
}
}
\description{
Optimize a regression-tree partitioning of the conditional outcome space.
}
\examples{
# ldag
levels <- list(0:1, 0:1)
counts <- cbind(c(10, 20, 30, 40), rep(10, 4))
optimize_partition_ldag(counts, levels, ess = 1, verbose = T)

# mixed cardinality
levels <- list(0:1, 0:2)
counts <- cbind(10, 1, c(10, 10, 10**2, 10**2, 10**3, 10**3))
res <- optimize_partition_ldag(counts, levels, ess = ess, verbose = T)
cbind(expand.grid(levels), n = counts, part = unlist_partition(res$partition))

# binary tree
levels <- list(0:1, 0:1)
counts <- cbind(c(10, 1000, 1000, 1000), rep(1, 4))
ess <- 1
res <- optimize_partition_tree(counts, levels, ess = ess, verbose = TRUE)
res$partition 
cbind(counts, unlist_partition(res$partition))


# mixed cardinality
levels <- list(0:1, 0:2)
counts <- cbind(10, 1, c(10, 10, 10**2, 10**2, 10**3, 10**3))
res <- optimize_partition_tree(counts, levels, ess = ess, verbose = T)
cbind(expand.grid(levels), n = counts, part = unlist_partition(res$partition))

# compare returned scores with score of partition
gr <- unlist_partition(res$partition)
tmp <- rowsum(counts, gr)
scores <- famscore_bdeu_byrow(tmp, ess, r = ncol(counts), q = nrow(counts), s = lengths(res$partition))
all.equal(scores, res$scores)
}
